// src/interfaces/IRepo.ts
export interface IRepo {
    avatar_url: string;
    name: string;
    login: string;
    watchers: number;
    stargazers_count: number;
    public_gists: number;
    subscribers_count: number;
    company: string;
    language: string;
    email: string;
    description: string;
    html_url: string;
}